+++ server/index.dev.js
@@ -133,6 +133,7 @@ export const createServer = (config) => {
                 <meta charSet="utf-8">
                 <meta httpEquiv="X-UA-Compatible" content="IE=edge">
                 ${head.title.toString()}
+                <meta http-equiv="Content-Security-Policy" content="block-all-mixed-content">
                 <meta name="viewport" content="width=device-width, initial-scale=1">
                 <link rel="shortcut icon" href="/favicon.ico">
                 ${head.meta.toString().replace(regexp, '')}