+++ CHANGELOG.md
@@ -10,13 +10,13 @@
 
 ### Changed
 - 優化 webpack bundle @esbb48
-  - 使用 CDN
+  - client 端使用 CDN 取代套件，減少壓縮時間
   - 移除不需要載入的元件（世大運）、Router（帳號密碼註冊登入）
   - 把共用的套件加到 Vendor
-  - 將 server 分成 prod 和 dev 兩種（避免在 build 的時候載入 dev 的 lib）
-  - 因應第四項調整 script
+  - 將 server 分成 prod 和 dev 兩種（避免在 build 的時候載入 dev 的 lib）並調整 script
   - 把 dev 的 lib 移到 devDependence
   - 使用 happypack 套件併發 build 的時間
+  - 壓縮 onead_ir 的 js 檔案，並移到本地端
 
 ## 1.0.24 - 2017-08-15
 ### Changed