+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
 ## [Unreleased]
+
+## 1.0.25 - 2017-08-24
 ### Note
 - `npm install happypack`